Ensure unhandled exceptions do not terminate shell, add throws matchers to expect (#66)
In the flutter_tester, unhandled expcetions will assert in the shell, causing the test to stop. Use a Zone level override of handleUncaughtError to instead forward it to logging to prevent the test runner from stopping.

Adds throws and throwsAsync top level functions to package:expect. These function similarly to the package:test integrated AsyncMatcher, but without the need to add callbacks to the test zone.