Adds a JSON manifest to cache test descriptions (#80)
Speeds up subsequent tester runs by caching the parsed test representation. On the flutter_tools general shard, this reduced a 200 ms operation to about 30 ms. There is a potential edge case that a large test manifest might be slower to parse than re-parsing a single file if the test file count is only 1. Consider an optimization in the future where all caches are elided for small test runs.

fixes #79