Add automatic concurrency (-j) level detection (#106)
So that it does not need to be specified each time. Make a trade-off between the number of test cases and the number of cores, preferring to stay in between the two numbers. This attempts to account for the fact that spawning a test runner has a non-zero cost, though a single test suite might not be the best measurement due to the package:test compatibility mode.